WIP splitdata for Neural Network
This is some code to split up the data between a training set and testing set. Then the goal was to validate it to get false positive/negative and true positives/negatives and make a stacked bar graph. Not fully working since onehot isn't lineing up with the split yet